+++ src/terra-dev-site/doc/application/Demo.4.doc.jsx
@@ -1,0 +1,3 @@
+import DemoAppIndex from './demo/DemoAppIndex';
+
+export default DemoAppIndex;